IRelationshipServicesAsync deleted functions

+++ GadevangTennisklub2025/Pages/TennisField/DeleteCourt.cshtml.cs
@@ -6,13 +6,15 @@ namespace GadevangTennisklub2025.Pages.TennisField
 {
     public class DeleteCourtModel : PageModel
     {
+        private IBookingServiceAsync _bookingService;
         private ICourtService _courtService;
         [BindProperty]
         public Models.TennisField Court { get; set; }
         [BindProperty] public bool Confirm { get; set; }
         public string MessageError { get; set; }
-        public DeleteCourtModel(ICourtService courtService)
+        public DeleteCourtModel(ICourtService courtService, IBookingServiceAsync bookingService)
         {
+            _bookingService = bookingService;
             _courtService = courtService;
         }
         public async Task<IActionResult> OnGetAsync(int deleteId)
@@ -33,6 +35,17 @@ namespace GadevangTennisklub2025.Pages.TennisField
                 MessageError = $"Husk at klikke på konfirmation";
                 return Page();
             }
+
+            var allBookings = await _bookingService.GetAllBookings();
+
+            bool courtHasBookings = allBookings.Any(b => b.Court_Id == Court.CourtId);
+
+            if (courtHasBookings)
+            {
+                MessageError = $"Kan ikke slette banen, da den har eksisterende bookinger.";
+                return Page();
+            }
+
             await _courtService.DeleteCourtAsync(Court.CourtId);
             return RedirectToPage("ShowCourts");
         }
